Removed tcolorbox pkg, \pagecolor, renamed \VK@ to \uis@
This commit removes the need for the tcolorbox package and the
corresponding \accessbox macro that used it. This is replaced by
the \colorbox supported by xcolor.

The \pagecolor renewcommand is also removed since we now have
a better approach to select colors, and don't need to read the
current page color anymore.

Finally, we also rename the \def\VK@ commands to \def\uis@ since
I don't know what \VK means... The naming was based on legacy
commands from the old his-brevmal.sty.